added the creation of the final shape
this should conclude the Internal Logic of the Cell, or at least the Graphics-related one. The Final Shape is a polygon built using the Base Shape and the semi-circles around it.
At worst, a Final Shape polygon would have 11524 points, hopefully it won' t cause any performance issues when there will be tens of cells drawn.