changed some naming stuff + added two new modules
I've changed the name of some modules(like "Cell" to "Cells") in order to avoid naming confusions(the class Cell had the same name as the module Cell and the code looked really weird). Also, I've slightly changed the code of the Cell class, by moving all the operations related to the MLCellApp(like adding the Final Item to the sene) in the code of the MLCellApp.
The Walls module contains the code for the Wall class, which corresponds to the walls that mark the boundaries of the graphics scene.
The utils module will contain a bunch of useful functions that aren't necessarily related to one class or another(like finding the class of an object etc.).