added pseudo-collision detection for cells + fixed spawning points for cells
The first version of cells collision is ready, it is not exactly perfect, but it is substantially faster than the Qt collision detection and is probably enough for our project.
The algorithm for choosing a suitable spawning point is fairly easy, it simply looks, for each cell, at all the other cells and sees which positions are occupied(ie positions on which the current cell cannot be drawn on without avoiding a collision) and randomly chooses one of the unoccupied positions. I will clarify this algorithm in the wiki.