optimized cell collision detection
Instead of checking collision with the whole otherArea rect in the vicinity function, the program is now only checking the rect's edges. Also made the time during which the cell moves in the opposite direction, after a collision, smaller, by reducing it from 1000 ms to 250 ms. It works fairly fast with 40 cells and the collision detection is, most of the time, ok.