added algae
I've just added the Alga class. The algae are small, greenish circles on the map. They spawn in Blooms, which are based on real lige Algae Blooms and represent two spawn-squares on the map. I've implemented some sort of collision with Cells, but it could use some improvements.
The number of Algae spawn is equal to half of the number of Cells on the map.