added smart eating + fixed some typos + some other stuff
Yet another huge commit. The most important addition of this commit is, by far, the "smart" food detection that cells now possess - now cells, based on their food preference, can choose what to eat, instead of the random though process they used to have.
Hunger was also implemented, although so far it only serves as a ticking doom clock for the cell, its only influence, besides killing the cell when it reaches 0, is that the cell won't look for food if its hunger-clock is at over 80%(if it isnt "hungry").
The "move" method was renamed to "updateLoop", since that's what it pretty much does, movement is only one of the thing it handles.
Many typos, mostly in old comments, were also fixed.
A "die" method was added to both cells and algae, which handles their death. Due to the weird way names work in Python, there has been added an if conditional in the updateLoop method, which checks if this cell should be dead, in which case it stops the function. This implies that, sometimes, although a cell should be dead, there are still names in the program that point to it. Not sure how to address this problem.