fixed some bugs + added algae spread options
The algae spread options, as their name suggests, model the way in which the algae will be spread on the map, meaning that they can either be concentrated in two smaller blooms(the "regular spread") or in two bigger blooms that encompass almost the entire map( the "full map spread"). This commit has backward compatibility, meaning that older saves that did not have this option will still work with this version.