added more graphs + greater mutation rate + fixed some bugs
I've added a bunch of extra graphs, mainly for showing the average values for individual generations, as opposed to the simulation average.
I've also changed the mutation rate, now the mutation rate for food preference is greater than before. It is chosen between -20 and 20 and the more carnivorous(or herbivorous) a cell is, the less likely it is to mutate its food preference towards the said preference(ie very carnivorous cells now have a greater chance to become less carnivorous after a mutation).
I've also fixed some bugs, mainly related to pausing and unpausing the game when saving\loading a sim.
The whole commit is backwards compatible,